# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Body of every non-2xx response (ADR 0006).",
        "properties": {
          "code": {
            "description": "Stable, machine-readable identifier in `snake_case`. Branch on this, never on `detail`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable message. Safe to show; never contains internals.",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present only for request-validation failures: one entry per offending field.",
            "title": "Errors"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FieldError": {
        "description": "One offending field of a request that failed validation.",
        "properties": {
          "loc": {
            "description": "Path to the field, starting with its location (`path`, `query`, `body`, ...).",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "message": {
            "description": "Human-readable reason the value was rejected.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "Machine-readable identifier of the violated rule (e.g. `less_than_equal`).",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "message",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "OperationCancelResponse": {
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          }
        },
        "required": [
          "operation_id",
          "kind",
          "status",
          "cancelled_at"
        ],
        "title": "OperationCancelResponse",
        "type": "object"
      },
      "OperationKind": {
        "enum": [
          "tutor_activate",
          "student_activate",
          "student_stash",
          "student_pop",
          "student_deactivate",
          "tutor_deactivate"
        ],
        "title": "OperationKind",
        "type": "string"
      },
      "OperationPage": {
        "description": "A page of operations plus the total match count, for offset pagination.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OperationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "OperationPage",
        "type": "object"
      },
      "OperationResponse": {
        "description": "Full read view of a single operation, including the two-phase ``plan``.",
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "failed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          },
          "reserved_archive_category_channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserved Archive Category Channel Id"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "subject_discord_id": {
            "title": "Subject Discord Id",
            "type": "integer"
          },
          "tutor_discord_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutor Discord Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "kind",
          "status",
          "guild_id",
          "subject_discord_id",
          "tutor_discord_id",
          "reserved_archive_category_channel_id",
          "plan",
          "expires_at",
          "committed_at",
          "cancelled_at",
          "failed_at",
          "last_error",
          "created_at",
          "updated_at"
        ],
        "title": "OperationResponse",
        "type": "object"
      },
      "OperationStatus": {
        "enum": [
          "prepared",
          "committed",
          "expired",
          "failed",
          "cancelled"
        ],
        "title": "OperationStatus",
        "type": "string"
      },
      "OperationSummary": {
        "description": "List view of an operation. Omits the heavy ``plan`` -- fetch a single operation by id for that.",
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "failed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "reserved_archive_category_channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserved Archive Category Channel Id"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "subject_discord_id": {
            "title": "Subject Discord Id",
            "type": "integer"
          },
          "tutor_discord_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutor Discord Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "kind",
          "status",
          "guild_id",
          "subject_discord_id",
          "tutor_discord_id",
          "reserved_archive_category_channel_id",
          "expires_at",
          "committed_at",
          "cancelled_at",
          "failed_at",
          "last_error",
          "created_at",
          "updated_at"
        ],
        "title": "OperationSummary",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2ClientCredentialsBearer": {
        "flows": {
          "clientCredentials": {
            "scopes": {
              "auth:clients:manage": "Manage application clients.",
              "bot:read": "Read bot API surface.",
              "bot:write": "Write bot API surface."
            },
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend of the skill-platform",
    "title": "skillforge",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bot/operations": {
      "get": {
        "description": "List operations by subject / status / kind, newest first, for post-restart reconciliation.\n\nAll filters are optional and AND-combined; the subject is the pair\n(`guild_id`, `subject_discord_id`). List items omit the `plan` -- read a single operation by id\nfor its full plan.",
        "operationId": "bot_list_operations",
        "parameters": [
          {
            "in": "query",
            "name": "guild_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Guild Id"
            }
          },
          {
            "in": "query",
            "name": "subject_discord_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Discord Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperationKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:read"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation_error",
                  "detail": "Request validation failed",
                  "errors": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "message": "Input should be less than or equal to 100",
                      "type": "less_than_equal"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:read"
            ]
          }
        ],
        "summary": "List Operations Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
    "/api/v1/bot/operations/{operation_id}": {
      "get": {
        "description": "Read a single operation by id, including its two-phase `plan`.",
        "operationId": "bot_read_operation",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:read"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_found": {
                    "value": {
                      "code": "operation_not_found",
                      "detail": "Operation not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation not found"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation_error",
                  "detail": "Request validation failed",
                  "errors": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "message": "Input should be less than or equal to 100",
                      "type": "less_than_equal"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:read"
            ]
          }
        ],
        "summary": "Read Operation",
        "tags": [
          "bot"
        ]
      }
    },
    "/api/v1/bot/operations/{operation_id}/cancel": {
      "post": {
        "description": "Explicitly cancel a PREPARED operation, releasing its reservation immediately.\n\nIdempotent on an already-cancelled operation; a non-cancellable state (committed, expired,\nfailed) returns 409 and an unknown id returns 404.",
        "operationId": "bot_cancel_operation",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_found": {
                    "value": {
                      "code": "operation_not_found",
                      "detail": "Operation not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation not found"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_pending": {
                    "value": {
                      "code": "operation_not_pending",
                      "detail": "Operation is not in a prepared state (already committed, failed, or expired)"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation is not in a prepared state (already committed, failed, or expired)"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation_error",
                  "detail": "Request validation failed",
                  "errors": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "message": "Input should be less than or equal to 100",
                      "type": "less_than_equal"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Cancel Operation Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
...
  },
...
}